pyproject.toml
[tool.poetry]
name = "event-detection"
version = "0.1.0"
description = "Event detection algorithms for NILM"
authors = ["Emilio J. Palacios-Garcia <emiliojose.palaciosgarcia@kuleuven.be>"]
readme = "README.md"
packages = [{include = "event_detection"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
paho-mqtt = "^1.6.1"
pyyaml = "^6.0"
numpy = "^1.25.0"
scipy = "^1.11.1"
psycopg2-binary = "^2.9.6"
pydantic = "^2.1.1"
pydantic-settings = "^2.0.2"
pandas = "^2.1.4"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
jupyterlab = "^4.0.4"
ipympl = "^0.9.3"
nbdime = "^3.2.1"
jupyterlab-git = "^0.43.0"
matplotlib = "^3.7.2"
nilmtk = {git = "https://github.com/ejpalacios/nilmtk-base.git"}

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
testpaths = [
    "tests/"
]

[tool.coverage.run]
omit = ["*/test*", "*__init__*"]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "scipy.*",
    "matplotlib.*",
    "event_detection.*"
]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
